Update: Create a base class for actions, linting errors
