+++ packages/dashboards/addon/components/dashboard-actions.ts
@@ -1,16 +1,14 @@
 /**
  * Copyright 2021, Yahoo Holdings Inc.
  * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
+ *
+ * Usage:
+ *   <DashboardActions
+ *      @item={{this.dashboard}}
+ *      @index={{this.index}}
+ *   />
  */
-import NaviActionList from 'navi-reports/components/navi-action-list';
-import { action } from '@ember/object';
+import NaviBaseActionList from 'navi-reports/components/navi-base-action-list';
+import type DashboardModel from 'navi-core/models/dashboard';
 
-export default class DashboardActionList extends NaviActionList {
-  /**
-   * @override
-   */
-  @action
-  isItemValid() {
-    return this.args.item.validations.isTruelyValid;
-  }
-}
+export default class DashboardActionList extends NaviBaseActionList<DashboardModel> {}
