+++ packages/reports/app/components/navi-base-action-list.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-reports/components/navi-base-action-list';
